clone-detect: share the type3/pipeline test helper (#3909)
The Type-3 and sequence "no self-overlap" tests were 25-line duplicates
(T3 0.95): same fixture-scan-assert shell, differing only in fixture,
config, and clone kind. Move that shell into the test-support crate as
`assert_single_file_has_no_overlapping_fragments` (gated behind an
optional `detect` feature so the scanner keeps reusing the filesystem
fixtures without pulling in clone-detect), and drive both cases from one
table-driven integration test. The now-unused `assert_no_overlapping_fragments`
helper is removed.

duplicated_lines 2216 -> 2191; type3_groups 80 -> 79; global gate PASS.

AI-generated with [Claude Code](https://claude.com/claude-code) (model opus)
