+++ packages/code/clone-detect/detect/src/tests/helpers.rs
@@ -3,7 +3,7 @@ use std::path::PathBuf;
 use clone_scanner::Config;
 use tempfile::TempDir;
 
-use crate::{DetectConfig, DetectionResult, Kind, instances};
+use crate::{DetectConfig, DetectionResult, instances};
 
 pub fn create_temp_file(dir: &TempDir, name: &str, content: &str) -> PathBuf {
     clone_test_support::write_file(dir.path(), name, content)
@@ -19,26 +19,4 @@ pub fn scan_and_run(dir: &TempDir, detect_config: &DetectConfig) -> DetectionRes
     let scanner = clone_scanner::Scanner::new(test_scan_config());
     let scan = scanner.directory(dir.path()).unwrap();
     instances(&scan, detect_config)
-}
-
-pub fn assert_no_overlapping_fragments(result: &DetectionResult, selected: impl Fn(&Kind) -> bool) {
-    for group in result
-        .instances
-        .iter()
-        .filter(|group| selected(&group.clone_type))
-    {
-        for (index, left) in group.fragments.iter().enumerate() {
-            for right in group.fragments.iter().skip(index + 1) {
-                let same_file = left.file == right.file;
-                let left_starts_before_right_ends = left.byte_range.start < right.byte_range.end;
-                let right_starts_before_left_ends = right.byte_range.start < left.byte_range.end;
-                let overlaps =
-                    same_file && left_starts_before_right_ends && right_starts_before_left_ends;
-                assert!(
-                    !overlaps,
-                    "clone group compared overlapping fragments: {group:?}"
-                );
-            }
-        }
-    }
 }
